# Optimize base64 decode AVX512 kernel for small input sizes