Optimize the base64 decode AVX512 routine for <64 bytes input and use Masked AVX512 instructions to process most of the <64 bytes tail.
Before the change:
```
 current system detected as icelake.
# loading files: .
# volume: 60 bytes
# max length: 60 bytes
# number of inputs: 1
# decode
memcpy                                   :   3.74 GB/s  17.75 GHz   4.75 c/b   4.15 i/b   0.87 i/c
libbase64                                :   1.98 GB/s  12.10 GHz   6.10 c/b  13.55 i/b   2.22 i/c
libbase64_space_decode                   :   1.03 GB/s   8.98 GHz   8.71 c/b  19.48 i/b   2.24 i/c
openssl3.3.x                             :   0.60 GB/s   7.58 GHz  12.63 c/b  45.42 i/b   3.60 i/c
node                                     :   1.65 GB/s  10.99 GHz   6.68 c/b  16.45 i/b   2.46 i/c
simdutf::icelake                         :   1.59 GB/s  10.17 GHz   6.40 c/b  12.62 i/b   1.97 i/c
simdutf::haswell                         :   1.54 GB/s  10.04 GHz   6.54 c/b  12.88 i/b   1.97 i/c
simdutf::westmere                        :   1.56 GB/s  10.14 GHz   6.49 c/b  12.70 i/b   1.96 i/c
simdutf::fallback                        :   1.96 GB/s  11.21 GHz   5.73 c/b  11.48 i/b   2.00 i/c
```
After the change:
```
# current system detected as icelake.
# loading files: .
# volume: 60 bytes
# max length: 60 bytes
# number of inputs: 1
# decode
memcpy                                   :   3.75 GB/s  17.65 GHz   4.71 c/b   4.15 i/b   0.88 i/c
libbase64                                :   2.03 GB/s  12.17 GHz   6.00 c/b  13.55 i/b   2.26 i/c
libbase64_space_decode                   :   1.08 GB/s   9.09 GHz   8.44 c/b  19.48 i/b   2.31 i/c
openssl3.3.x                             :   0.60 GB/s   7.59 GHz  12.76 c/b  45.42 i/b   3.56 i/c
node                                     :   1.63 GB/s  10.95 GHz   6.72 c/b  16.45 i/b   2.45 i/c
simdutf::icelake                         :   2.39 GB/s  12.46 GHz   5.22 c/b   6.30 i/b   1.21 i/c
simdutf::haswell                         :   1.53 GB/s  10.03 GHz   6.54 c/b  12.88 i/b   1.97 i/c
simdutf::westmere                        :   1.54 GB/s  10.03 GHz   6.51 c/b  12.70 i/b   1.95 i/c
simdutf::fallback                        :   1.82 GB/s  10.59 GHz   5.81 c/b  11.48 i/b   1.98 i/c
```