@lemire 2024-12-01
Still on my todo