@lemire 2024-12-10
Great work. Merging.